… go/vt/sqlparser/sql.go