For a high QPS application, the sqlparser generates a ton of garbage.
```
622GB    26.33% 60.87%   880.50GB 37.28% vitess.io/vitess/go/vt/sqlparser.yyParse
212.57GB  9.00% 69.87%   258.49GB 10.94% vitess.io/vitess/go/vt/sqlparser.(*yyParserImpl).Parse
```

Most of this, per the heap profile is allocations of new `yyParserImpl`
objects for every call to `yyParse`. This diff uses `sync.Pool` to
mitigate the total allocation overhead. Results:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkNormalize-4          2540          2641          +3.98%
BenchmarkParse1-4             18269         13923         -23.79%
BenchmarkParse2-4             46703         41721         -10.67%
BenchmarkParse2Parallel-4     22246         20290         -8.79%
BenchmarkParse3-4             4064743       4151772       +2.14%

benchmark                     old allocs     new allocs     delta
BenchmarkNormalize-4          27             27             +0.00%
BenchmarkParse1-4             75             74             -1.33%
BenchmarkParse2-4             264            263            -0.38%
BenchmarkParse2Parallel-4     176            175            -0.57%
BenchmarkParse3-4             360            361            +0.28%

benchmark                     old bytes     new bytes     delta
BenchmarkNormalize-4          821           821           +0.00%
BenchmarkParse1-4             22776         2304          -89.88%
BenchmarkParse2-4             28352         7875          -72.22%
BenchmarkParse2Parallel-4     25712         5234          -79.64%
BenchmarkParse3-4             6352082       6336288       -0.25%
```

Resolves: https://github.com/vitessio/vitess/issues/4174

Signed-off-by: Daniel Tahara <tahara@dropbox.com>