@danieltahara 2018-09-03
Getting some error on ParseNext, but only for one of the tests ```➜  sqlparser git:(parser-pool) ✗ go test
--- FAIL: TestIgnoreSpecialComments (0.00s)
        parse_next_test.go:77: got select 1 from dual want select 2 from dual
```