@sougou 2018-09-04
I'm actually surprised you're getting any benefit. For large size allocations like in this case, the cost of zeroing the memory is significant. This is something you're doing in a loop whereas the runtime can do it as a sweep.

The variance is probably related to just the cost of obtaining the memory, which will vary depending on fragmentation level. Are you able to get consistent numbers?

One experiment you can try (instead of the loop) is to have a static zero-initialized yyParserImpl, and just assign that to the one you got from the pool, which is likely to zero it out more efficiently.