@danieltahara 2018-09-04
That's a good idea.

In any event, there are some issues with holding references to the "backing" stack. I need to sort those out before going forward