@danieltahara 2018-09-04
And to answer your other question, the numbers are consistent from run to run.